Move page logic to feature folder
