+++ idea/frontend/src/features/voucher/index.ts
@@ -1,6 +1,7 @@
 import { IssueVoucher, ProgramVoucherSelect, CodeVoucherSelect, VoucherTable, VoucherBadge, VoucherCard } from './ui';
-
 import VoucherCardPlaceholder from './assets/voucher-card-placeholder.svg?react';
+import { useVouchers, useVoucherFilters } from './hooks';
+import { Voucher } from './types';
 
 export {
   IssueVoucher,
@@ -10,4 +11,8 @@ export {
   VoucherBadge,
   VoucherCard,
   VoucherCardPlaceholder,
+  useVouchers,
+  useVoucherFilters,
 };
+
+export type { Voucher };
