+++ idea/frontend/src/features/voucher/types.ts
@@ -1,3 +1,4 @@
+import { HexString } from '@gear-js/api';
 import { FIELD_NAME, VOUCHER_TYPE } from './consts';
 
 type VoucherType = typeof VOUCHER_TYPE[keyof typeof VOUCHER_TYPE];
@@ -8,4 +9,36 @@ type Values = {
   [FIELD_NAME.DURATION]: string;
 };
 
-export type { VoucherType, Values };
+type Voucher = {
+  id: HexString;
+  owner: HexString;
+  spender: HexString;
+  amount: string;
+  balance: string;
+  programs: HexString[];
+  codeUploading: boolean;
+  expiryAtBlock: string;
+  expiryAt: string;
+  issuedAtBlock: string;
+  issuedAt: string;
+  updatedAtBlock: string;
+  updatedAt: string;
+  isDeclined: boolean;
+};
+
+type VouchersParams = {
+  limit: number;
+  offset: number;
+  id: HexString;
+  owner: string;
+  spender: string;
+  declined: boolean;
+  expired: boolean;
+};
+
+type VouchersResponse = {
+  vouchers: Voucher[];
+  count: number;
+};
+
+export type { VoucherType, Values, Voucher, VouchersParams, VouchersResponse };
