+++ idea/frontend/src/features/voucher/utils.ts
@@ -1,3 +1,7 @@
+import { DEFAULT_LIMIT, VOUCHERS_API_URL } from '@/shared/config';
+
+import { VouchersResponse } from './types';
+
 const MULTIPLIER = { MS: 1000, S: 60, M: 60, H: 24, D: 30 };
 
 function getMilliseconds(value: number, unit: 'hour' | 'day' | 'month') {
@@ -28,4 +32,29 @@ function getTime(ms: number) {
   return result.trim();
 }
 
-export { getPluralizedUnit, getMilliseconds, getTime };
+const fetchWithGuard = async <T extends object>(...args: Parameters<typeof fetch>) => {
+  const response = await fetch(...args);
+
+  if (!response.ok) throw new Error(response.statusText);
+
+  return response.json() as T;
+};
+
+const getVouchers = (params: { offset: number; limit: number; id: string }) => {
+  const url = `${VOUCHERS_API_URL}/vouchers`;
+  const method = 'POST';
+  const body = JSON.stringify(params);
+  const headers = { 'Content-Type': 'application/json' };
+
+  return fetchWithGuard<VouchersResponse>(url, { method, body, headers });
+};
+
+const getNextPageParam = (lastPage: VouchersResponse, allPages: VouchersResponse[]) => {
+  const totalCount = lastPage.count;
+  const lastPageCount = lastPage.vouchers.length;
+  const fetchedCount = (allPages.length - 1) * DEFAULT_LIMIT + lastPageCount;
+
+  return fetchedCount < totalCount ? fetchedCount : undefined;
+};
+
+export { getPluralizedUnit, getMilliseconds, getTime, getVouchers, getNextPageParam };
