--- idea/frontend/src/pages/vouchers/consts.ts
@@ -1,3 +1,0 @@
-const PAGE_SIZE = 20;
-
-export { PAGE_SIZE };
