--- idea/frontend/src/pages/vouchers/hooks/index.ts
@@ -1,4 +1,0 @@
-import { useVoucherFilters } from './use-voucher-filters';
-import { useVouchers } from './use-vouchers';
-
-export { useVoucherFilters, useVouchers };
