+++ idea/frontend/src/shared/ui/form/index.ts
@@ -4,5 +4,6 @@ import { ValueField } from './value-field';
 import { Checkbox } from './checkbox';
 import { Radio } from './radio';
 import { Select } from './select';
+import { SearchForm } from './search-form';
 
-export { formStyles, Input, ValueField, Checkbox, Radio, Select };
+export { formStyles, Input, ValueField, Checkbox, Radio, Select, SearchForm };
