+++ idea/frontend/src/shared/ui/form/search-form/index.ts
