+++ idea/frontend/src/shared/ui/index.ts
@@ -1,6 +1,20 @@
 import { withAccount } from './with-account';
 import { LabeledCheckbox } from './labeled-checkbox';
-import { Input, ValueField, Checkbox, Radio, Select } from './form';
+import { Input, ValueField, Checkbox, Radio, Select, SearchForm } from './form';
 import { withDeprecatedFallback } from './with-deprecated-fallback';
+import { List } from './list';
+import { Skeleton } from './skeleton';
 
-export { LabeledCheckbox, withAccount, Input, ValueField, Checkbox, Radio, Select, withDeprecatedFallback };
+export {
+  LabeledCheckbox,
+  withAccount,
+  Input,
+  ValueField,
+  Checkbox,
+  Radio,
+  Select,
+  withDeprecatedFallback,
+  List,
+  Skeleton,
+  SearchForm,
+};
