+++ idea/frontend/src/shared/ui/list/index.ts
