+++ idea/frontend/src/shared/ui/list/list.module.scss
