+++ idea/frontend/src/shared/ui/list/list.tsx
