+++ idea/frontend/src/shared/ui/list/observer.tsx
