+++ idea/frontend/src/shared/ui/skeleton/index.ts
