+++ idea/frontend/src/shared/ui/skeleton/skeleton.module.scss
