+++ idea/frontend/src/shared/ui/skeleton/skeleton.tsx
